Introduce promise_test().
promise_test() is similar in character to async_test() but intends to
reduce the boilerplate of writing tests for code that return promises.

It requires a function object as the first parameter. This function,
when called with a test object as its only parameter is expected to
return a thenable (a Promise or an object that supports .then and .catch
methods).

When the returned thenable fulfills, the test is considered done. If the
thenable rejects, the test fails in a manner similar to having been
terminated with an unhandled exception.