Support synthesizing tests from test results.
Introduce synthesize_tests(). This method when called with an array of
test results and a harness status creates a set of synthesized tests
that mirror the provided results.

A service worker or a web worker can run full fidelity tests and then
pass the results over to a client document. This client document can
then call synthesize_tests() to replicate the test run from the worker.
Any reporting infrastructure that's already in place can then handle the
worker's test results in the same manner as handling tests run directly
on the client document.