Support using testharness.js from a worker.
This change introduces an abstraction for aspects of testharness.js that
require a DOM and adds two implementations of this abstraction:
WindowTestEnvironment and WorkerTestEnvironment. The latter is safe for
use within a web worker or a service worker.

Within a worker, the test harness can be imported via:

   importScripts('/resources/testharness.js');

Once imported, the worker can invoke tests using test(), async_test()
etc. with the following caveats:

- Test results won't be reported anywhere by default. A mechanism by which
  a client document can capture and expose test results from a worker
  will be added soon.

- Lacking common metadata usually provided via <title> and <meta>
  elements, each test much provide its own properties.

There should be no behavior change for regular testharness based tests.